requirements.txt
streamlit
pillow
pandas
plotly
google-generativeai
PyPDF2
langchain
